Add files via upload
little bug fix 
if parameter = 0 calculation are executed, variable type in main should be same (long) than method return

+++ SpeedConverter/src/Main.java
@@ -2,7 +2,7 @@ public class Main {
 
     public static void main(String[] args){
 
-        double s = SpeedConverter.toMilesPerHour(1.5);
+        long s = SpeedConverter.toMilesPerHour(1.5);
         System.out.println("first method " + s);
         SpeedConverter.printConversion(-2);
     }
